+++ CHANGELOG.md
@@ -234,3 +234,4 @@
 # 0.2.9
 * Define `Zipper` type family to walk datastructures with context
 * Define `Tap` datatype for context dependent values
+* Define type synonyms for `TU` and `UT` joint schemes